ml-service-python/requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML and AI libraries
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.0
transformers==4.35.0
sentencepiece==0.1.99

# Image and media processing
Pillow==10.0.0
opencv-python==4.8.0.76
librosa==0.10.1
numpy==1.24.3

# Additional utilities
requests==2.31.0
aiofiles==23.2.1
scikit-learn==1.3.0

# Optional: For better performance
accelerate==0.23.0
